# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Benchmark {
  /** Big-O time complexity, e.g. O(n log n) */
  timeComplexity: string;
  /** Big-O space complexity, e.g. O(n) */
  spaceComplexity: string;
  /** Estimated runtime in milliseconds for a representative input */
  estimatedTimeMs: number;
  /** Estimated peak memory usage in kilobytes */
  estimatedMemoryKb: number;
  /** Score 0-100 of how well this variant balances time and memory */
  balanceScore: number;
}

export interface Variant {
  id: string;
  /** e.g. "sorting", "two-pointers", "dynamic-programming" */
  technique: string;
  /** Human-readable technique label */
  techniqueLabel: string;
  title: string;
  /** Short blog-style description of the approach */
  summary: string;
  /** Detailed explanation of how this variant works */
  explanation: string;
  language: string;
  code: string;
  benchmark: Benchmark;
  createdAt: string;
}

export interface AnalyzeRequest {
  /** @minLength 1 */
  code: string;
  /** Optional language hint */
  hintLanguage?: string;
}

export interface AnalyzeResponse {
  detectedLanguage: string;
  originalCode: string;
  fixedCode: string;
  errorsFound: string[];
  /** Short summary of what the input program does */
  problemSummary: string;
  variants: Variant[];
  /** Id of the variant that best balances time and memory */
  recommendedVariantId: string;
}

export interface ConvertRequest {
  variantId: string;
  /** e.g. python, java, c, cpp, javascript, ruby, rust, go, swift, kotlin */
  targetLanguage: string;
}

export interface ConvertResponse {
  variantId: string;
  targetLanguage: string;
  code: string;
  /** Notes about the conversion (idiomatic adjustments, etc.) */
  notes: string;
}

export interface BlogTechniqueGroup {
  technique: string;
  techniqueLabel: string;
  variants: Variant[];
}

export interface BlogList {
  groups: BlogTechniqueGroup[];
  totalVariants: number;
}

export interface TechniqueInfo {
  slug: string;
  label: string;
  description: string;
}

export interface TechniqueList {
  techniques: TechniqueInfo[];
}