# lib/api-zod/src/generated/types/analyzeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeRequest {
  /** @minLength 1 */
  code: string;
  /** Optional language hint */
  hintLanguage?: string;
}