# lib/api-zod/src/generated/types/analyzeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Variant } from "./variant";

export interface AnalyzeResponse {
  detectedLanguage: string;
  originalCode: string;
  fixedCode: string;
  errorsFound: string[];
  /** Short summary of what the input program does */
  problemSummary: string;
  variants: Variant[];
  /** Id of the variant that best balances time and memory */
  recommendedVariantId: string;
}