# lib/api-zod/src/generated/types/benchmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Benchmark {
  /** Big-O time complexity, e.g. O(n log n) */
  timeComplexity: string;
  /** Big-O space complexity, e.g. O(n) */
  spaceComplexity: string;
  /** Estimated runtime in milliseconds for a representative input */
  estimatedTimeMs: number;
  /** Estimated peak memory usage in kilobytes */
  estimatedMemoryKb: number;
  /** Score 0-100 of how well this variant balances time and memory */
  balanceScore: number;
}