# lib/api-zod/src/generated/types/blogList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogTechniqueGroup } from "./blogTechniqueGroup";

export interface BlogList {
  groups: BlogTechniqueGroup[];
  totalVariants: number;
}