# lib/api-zod/src/generated/types/blogTechniqueGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Variant } from "./variant";

export interface BlogTechniqueGroup {
  technique: string;
  techniqueLabel: string;
  variants: Variant[];
}