# lib/api-zod/src/generated/types/convertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertRequest {
  variantId: string;
  /** e.g. python, java, c, cpp, javascript, ruby, rust, go, swift, kotlin */
  targetLanguage: string;
}