# lib/api-zod/src/generated/types/convertResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertResponse {
  variantId: string;
  targetLanguage: string;
  code: string;
  /** Notes about the conversion (idiomatic adjustments, etc.) */
  notes: string;
}