# lib/api-zod/src/generated/types/techniqueInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TechniqueInfo {
  slug: string;
  label: string;
  description: string;
}