# lib/api-zod/src/generated/types/techniqueList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TechniqueInfo } from "./techniqueInfo";

export interface TechniqueList {
  techniques: TechniqueInfo[];
}