# lib/api-zod/src/generated/types/variant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Benchmark } from "./benchmark";

export interface Variant {
  id: string;
  /** e.g. "sorting", "two-pointers", "dynamic-programming" */
  technique: string;
  /** Human-readable technique label */
  techniqueLabel: string;
  title: string;
  /** Short blog-style description of the approach */
  summary: string;
  /** Detailed explanation of how this variant works */
  explanation: string;
  language: string;
  code: string;
  benchmark: Benchmark;
  createdAt: string;
}